pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "amr_rl"
version = "0.1.0"
description = "Adversarially co-trained RL for antimicrobial resistance treatment"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=0.29",
    "numpy>=1.26",
    "torch>=2.2",
    "stable-baselines3>=2.2",
    "scipy>=1.12",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "tqdm>=4.66",
    "wandb>=0.16",
    "pyyaml>=6.0",
    "lifelines>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov", "ruff", "mypy"]
notebooks = ["jupyter", "ipywidgets"]

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"
